requirements.txt
# Core backend
django==4.2.6
djangorestframework==3.14.0
django-extensions==3.2.3
django-cors-headers==4.2.0
django-filter==23.3

# REST API
djangorestframework==3.14.0
drf-spectacular==0.26.5
# Cosmetic
django-bootstrap5==23.3

# Database
psycopg2-binary==2.9.9

# Webserver
gunicorn==21.2.0

# Celery
celery==5.3.4
msgpack-python==0.5.6
django-celery-results==2.5.1